Add delete alert dialog

+++ lib/components/exercise_card.dart
@@ -47,8 +47,26 @@ class _ExerciseCardState extends State<ExerciseCard> {
         subtitle: widget.exercise.description != null ? Text(widget.exercise.description!) :Text(""),
         trailing: IconButton(
             icon: const Icon(Icons.more_vert),
-            onPressed: (){
-              database.deleteExercise(widget.exercise);
+            onPressed: (){ showDialog(
+                context: context,
+                builder: (BuildContext context) => AlertDialog(
+                  title: const Text("Warnung"),
+                  content: const Text("Soll diener Eintrag geloescht werden?"),
+                  actions: <Widget>[
+                    TextButton(
+                      onPressed: () => Navigator.pop(context, 'Cancel'),
+                      child: const Text('Cancel'),
+                    ),
+                    TextButton(
+                      onPressed: () {
+                        database.deleteExercise(widget.exercise);
+                        Navigator.pop(context, 'Delete');
+                      },
+                      child: const Text('Delete'),
+                    ),
+                  ],
+            ));
+              ;
             }),
         onTap: () {
           Navigator.of(context).pushNamed(
